Revert unneeded changes
This reverts commit cc74de037ff4b8c785ebd2624ddcdb0bda6c1b17.
