+++ src/components/ShareExperience/TimeSalaryForm/index.js
@@ -324,7 +324,6 @@ const TimeSalaryForm = () => {
           hasOvertimeSalary={hasOvertimeSalary}
           isOvertimeSalaryLegal={isOvertimeSalaryLegal}
           hasCompensatoryDayoff={hasCompensatoryDayoff}
-          submitted={submitted}
         />
       </IconHeadingBlock>
